+++ packages/7715-permission-types/src/types.ts
@@ -191,7 +191,7 @@ export type TokenApprovalRevocationPermission = BasePermission & {
     erc20Approve: boolean;
     erc721Approve: boolean;
     erc721SetApprovalForAll: boolean;
-    permit2ApproveZero: boolean;
+    permit2Approve: boolean;
     permit2Lockdown: boolean;
     permit2InvalidateNonces: boolean;
   };