+++ packages/delegator-e2e/test/caveats/approvalRevocation.test.ts
@@ -87,7 +87,7 @@ test('maincase: Bob revokes an existing ERC20 approval', async () => {
         erc20Approve: true,
         erc721Approve: false,
         erc721SetApprovalForAll: false,
-        permit2ApproveZero: false,
+        permit2Approve: false,
         permit2Lockdown: false,
         permit2InvalidateNonces: false,
       })